ElReto/requirements.txt
opencv-python==4.8.1
numpy==1.24.3
google-generativeai==0.3.2
python-dotenv==1.0.0
pillow==10.1.0